Model Fixes
I moved all models from other files into the models folder.
I renamed dmJumpTo to JumpTo so that it will be easy for identification.
I also made corrections to the zuriBottomSheet dark mode coloring.
