+++ lib/models/drafted_item_model.dart
@@ -1,0 +1,9 @@
+class DraftedItem {
+  final String? username, message, time;
+
+  DraftedItem({
+    this.username,
+    this.message,
+    this.time,
+  });
+}
