+++ lib/models/message.dart
