+++ lib/ui/shared/bottom_sheets/zuri_chat_bottomsheet.dart
@@ -31,6 +31,7 @@ Future<dynamic> zuriChatBottomSheet({
     enableDrag: true,
     isDismissible: true,
     builder: (context) {
+      final bool _dark = Theme.of(context).brightness == Brightness.dark;
       return Container(
         width: double.infinity,
         height: MediaQuery.of(context).size.height * .40,
@@ -73,11 +74,13 @@ Future<dynamic> zuriChatBottomSheet({
             // ListTile(
             //   title: Text(
             //     MarkUnread,
-            //     style: AppTextStyle.darkGreySize16,
+            //     style: _dark
+            // ? AppTextStyle.whiteSize16
+            // : AppTextStyle.darkGreySize16,
             //   ),
             //   leading: SvgPicture.asset(
             //     Mark_Unread,
-            //     color: AppColors.darkGreyColor,
+            //     color: _dark ? AppColors.whiteColor : AppColors.darkGreyColor,
             //     width: 18,
             //     height: 18,
             //   ),
@@ -86,11 +89,13 @@ Future<dynamic> zuriChatBottomSheet({
             ListTile(
               title: Text(
                 RemindMe,
-                style: AppTextStyle.darkGreySize16,
+                style: _dark
+                    ? AppTextStyle.whiteSize16
+                    : AppTextStyle.darkGreySize16,
               ),
               leading: SvgPicture.asset(
                 Remind_Me,
-                color: AppColors.darkGreyColor,
+                color: _dark ? AppColors.whiteColor : AppColors.darkGreyColor,
                 width: 18,
                 height: 18,
               ),
@@ -101,11 +106,13 @@ Future<dynamic> zuriChatBottomSheet({
             ListTile(
               title: Text(
                 AddToSavedItems,
-                style: AppTextStyle.darkGreySize16,
+                style: _dark
+                    ? AppTextStyle.whiteSize16
+                    : AppTextStyle.darkGreySize16,
               ),
               leading: SvgPicture.asset(
                 Saved_Items,
-                color: AppColors.darkGreyColor,
+                color: _dark ? AppColors.whiteColor : AppColors.darkGreyColor,
                 width: 18,
                 height: 18,
               ),
@@ -115,11 +122,13 @@ Future<dynamic> zuriChatBottomSheet({
             // ListTile(
             //     title: Text(
             //       ReplyInThreads,
-            //       style: AppTextStyle.darkGreySize16,
+            //       style: _dark
+            // ? AppTextStyle.whiteSize16
+            // : AppTextStyle.darkGreySize16,
             //     ),
             //     leading: SvgPicture.asset(
             //       Reply_In_Thread,
-            //       color: AppColors.darkGreyColor,
+            //       color: _dark ? AppColors.whiteColor : AppColors.darkGreyColor,
             //       width: 18,
             //       height: 18,
             //     ),
@@ -127,11 +136,13 @@ Future<dynamic> zuriChatBottomSheet({
             // ListTile(
             //     title: Text(
             //       FollowThreadZuriChatBottomSheet,
-            //       style: AppTextStyle.darkGreySize16,
+            //       style: _dark
+            // ? AppTextStyle.whiteSize16
+            // : AppTextStyle.darkGreySize16,
             //     ),
             //     leading: SvgPicture.asset(
             //       Follow_Thread,
-            //       color: AppColors.darkGreyColor,
+            //       color: _dark ? AppColors.whiteColor : AppColors.darkGreyColor,
             //       width: 18,
             //       height: 18,
             //     ),
@@ -140,28 +151,34 @@ Future<dynamic> zuriChatBottomSheet({
             ListTile(
                 title: Text(
                   ShareMessage,
-                  style: AppTextStyle.darkGreySize16,
+                  style: _dark
+                      ? AppTextStyle.whiteSize16
+                      : AppTextStyle.darkGreySize16,
                 ),
                 leading: SvgPicture.asset(
                   Share_message,
-                  color: AppColors.darkGreyColor,
+                  color: _dark ? AppColors.whiteColor : AppColors.darkGreyColor,
                   width: 18,
                   height: 18,
                 ),
                 onTap: shareMessage),
             ListTile(
-                title:
-                    Text('Delete Message', style: AppTextStyle.darkGreySize16),
+                title: Text('Delete Message',
+                    style: _dark
+                        ? AppTextStyle.whiteSize16
+                        : AppTextStyle.darkGreySize16),
                 leading: const Icon(Icons.delete),
                 onTap: deleteMessage),
             // ListTile(
             //     title: Text(
             //       CopyLinkToMessage,
-            //       style: AppTextStyle.darkGreySize16,
+            //       style: _dark
+            // ? AppTextStyle.whiteSize16
+            // : AppTextStyle.darkGreySize16,
             //     ),
             //     leading: SvgPicture.asset(
             //       Copy_link_To_Message,
-            //       color: AppColors.darkGreyColor,
+            //       color: _dark ? AppColors.whiteColor : AppColors.darkGreyColor,
             //       width: 18,
             //       height: 18,
             //     ),
@@ -169,11 +186,13 @@ Future<dynamic> zuriChatBottomSheet({
             ListTile(
                 title: Text(
                   CopyTextZuriChatBottomSheet,
-                  style: AppTextStyle.darkGreySize16,
+                  style: _dark
+                      ? AppTextStyle.whiteSize16
+                      : AppTextStyle.darkGreySize16,
                 ),
                 leading: SvgPicture.asset(
                   Copy_Text,
-                  color: AppColors.darkGreyColor,
+                  color: _dark ? AppColors.whiteColor : AppColors.darkGreyColor,
                   width: 18,
                   height: 18,
                 ),
@@ -182,11 +201,13 @@ Future<dynamic> zuriChatBottomSheet({
             ListTile(
                 title: Text(
                   "${post?.pinned == true ? local!.unPinFrom : local!.pinTo} ${local.conversation}",
-                  style: AppTextStyle.darkGreySize16,
+                  style: _dark
+                      ? AppTextStyle.whiteSize16
+                      : AppTextStyle.darkGreySize16,
                 ),
                 leading: SvgPicture.asset(
                   pinned_message,
-                  color: AppColors.darkGreyColor,
+                  color: _dark ? AppColors.whiteColor : AppColors.darkGreyColor,
                   width: 18,
                   height: 18,
                 ),
