+++ lib/ui/view/drafted_messages_page/widgets/app/data.dart
@@ -1,14 +1,5 @@
+import 'package:zurichat/models/drafted_item_model.dart';
 import 'package:zurichat/utilities/constants/app_strings.dart';
-
-class DraftedItem {
-  final String? username, message, time;
-
-  DraftedItem({
-    this.username,
-    this.message,
-    this.time,
-  });
-}
 
 double mySize = 0.0;
 double kPadding = mySize * 0.025;
