+++ lib/ui/view/edit_profile/widget/edit_profile_body.dart
@@ -5,9 +5,9 @@ import 'package:zurichat/ui/shared/shared.dart';
 import 'package:stacked/stacked.dart';
 import '../edit_profile_viewmodel.dart';
 
-class Body extends ViewModelWidget<EditProfileViewModel> {
+class EditProfileBody extends ViewModelWidget<EditProfileViewModel> {
   final Size _size;
-  const Body({
+  const EditProfileBody({
     Key? key,
     required Size size,
   })  : _size = size,
