+++ lib/ui/view/file_search/file_search_view.dart
@@ -103,5 +103,3 @@ class _FileSearchViewState extends State<FileSearchView> {
     );
   }
 }
-
-//* Custom Widgets
